[Games] added: dummy window and database for games
