+++ xbmc.vcproj
@@ -1259,12 +1259,6 @@ bundler xbelogo\saveimage.rdf -o xbelogo\saveimage.xbx
 					RelativePath=".\xbmc\DatabaseManager.h">
 				</File>
 				<File
-					RelativePath=".\xbmc\ProgramDatabase.cpp">
-				</File>
-				<File
-					RelativePath=".\xbmc\ProgramDatabase.h">
-				</File>
-				<File
 					RelativePath=".\xbmc\TextureDatabase.cpp">
 				</File>
 				<File
@@ -1426,12 +1420,6 @@ bundler xbelogo\saveimage.rdf -o xbelogo\saveimage.xbx
 					RelativePath=".\xbmc\pictures\GUIViewStatePictures.h">
 				</File>
 				<File
-					RelativePath=".\xbmc\programs\GUIViewStatePrograms.cpp">
-				</File>
-				<File
-					RelativePath=".\xbmc\programs\GUIViewStatePrograms.h">
-				</File>
-				<File
 					RelativePath=".\xbmc\windows\GUIWindowFileManager.cpp">
 				</File>
 				<File
@@ -1460,9 +1448,6 @@ bundler xbelogo\saveimage.rdf -o xbelogo\saveimage.xbx
 				</File>
 				<File
 					RelativePath=".\xbmc\windows\GUIWindowPointer.cpp">
-				</File>
-				<File
-					RelativePath=".\xbmc\programs\GUIWindowPrograms.cpp">
 				</File>
 				<File
 					RelativePath=".\xbmc\windows\GUIWindowScreensaver.cpp">
@@ -4618,6 +4603,28 @@ bundler xbelogo\saveimage.rdf -o xbelogo\saveimage.xbx
 					RelativePath=".\xbmc\input\KeyboardLayoutManager.h">
 				</File>
 			</Filter>
+			<Filter
+				Name="programs"
+				Filter="">
+				<File
+					RelativePath=".\xbmc\programs\GUIViewStatePrograms.cpp">
+				</File>
+				<File
+					RelativePath=".\xbmc\programs\GUIViewStatePrograms.h">
+				</File>
+				<File
+					RelativePath=".\xbmc\programs\GUIWindowPrograms.cpp">
+				</File>
+				<File
+					RelativePath=".\xbmc\programs\GUIWindowPrograms.h">
+				</File>
+				<File
+					RelativePath=".\xbmc\programs\ProgramDatabase.cpp">
+				</File>
+				<File
+					RelativePath=".\xbmc\programs\ProgramDatabase.h">
+				</File>
+			</Filter>
 		</Filter>
 		<Filter
 			Name="Header Files"
@@ -4781,9 +4788,6 @@ bundler xbelogo\saveimage.rdf -o xbelogo\saveimage.xbx
 			</File>
 			<File
 				RelativePath=".\xbmc\windows\GUIWindowPointer.h">
-			</File>
-			<File
-				RelativePath=".\xbmc\programs\GUIWindowPrograms.h">
 			</File>
 			<File
 				RelativePath=".\xbmc\windows\GUIWindowScreensaver.h">
