+++ xbmc/DatabaseManager.cpp
@@ -22,8 +22,8 @@
 #include "utils/log.h"
 #include "addons/AddonDatabase.h"
 #include "view/ViewDatabase.h"
-#include "ProgramDatabase.h"
 #include "TextureDatabase.h"
+#include "programs/ProgramDatabase.h"
 #include "music/MusicDatabase.h"
 #include "video/VideoDatabase.h"
 #include "settings/AdvancedSettings.h"
