+++ xbmc/programs/ProgramDatabase.cpp
@@ -1,0 +1,56 @@
+/*
+ *  Copyright (C) 2025-2025 Team XBMC
+ *  This file is part of XBMC - https://github.com/antonic901/xbmc4xbox-redux
+ *
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *  See LICENSE.md for more information.
+ */
+
+#include "ProgramDatabase.h"
+
+#include "dbwrappers/dataset.h"
+#include "utils/log.h"
+#include "utils/StringUtils.h"
+
+using namespace dbiplus;
+
+CProgramDatabase::CProgramDatabase(void)
+{
+}
+
+CProgramDatabase::~CProgramDatabase(void)
+{}
+
+bool CProgramDatabase::Open()
+{
+  return CDatabase::Open();
+}
+
+void CProgramDatabase::CreateTables()
+{
+  CLog::Log(LOGINFO, "create path table");
+  m_pDS->exec("CREATE TABLE path (idPath integer primary key, strPath text, strContent text, strHash text, dateAdded text)");
+
+  CLog::Log(LOGINFO, "create program table");
+  std::string columns = "CREATE TABLE program (idProgram integer primary key, idPath integer";
+
+  for (int i = 0; i < PROGRAMDB_MAX_COLUMNS; i++)
+    columns += StringUtils::Format(",c%02d text", i);
+
+  columns += ")";
+  m_pDS->exec(columns);
+}
+
+void CProgramDatabase::CreateAnalytics()
+{
+  CLog::Log(LOGINFO, "%s - creating indicies", __FUNCTION__);
+  m_pDS->exec("CREATE INDEX ix_path ON path ( strPath(255) )");
+
+  m_pDS->exec("CREATE UNIQUE INDEX ix_program_file_1 ON program (idPath, idProgram)");
+  m_pDS->exec("CREATE UNIQUE INDEX ix_program_file_2 ON program (idProgram, idPath)");
+}
+
+int CProgramDatabase::GetSchemaVersion() const
+{
+  return 1;
+}
