+++ xbmc/programs/ProgramDatabase.h
@@ -1,0 +1,30 @@
+#pragma once
+/*
+ *  Copyright (C) 2025-2025 Team XBMC
+ *  This file is part of XBMC - https://github.com/antonic901/xbmc4xbox-redux
+ *
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *  See LICENSE.md for more information.
+ */
+
+#include <string>
+
+#include "dbwrappers/Database.h"
+
+#define PROGRAMDB_MAX_COLUMNS 24
+
+class CProgramDatabase : public CDatabase
+{
+public:
+  CProgramDatabase(void);
+  virtual ~CProgramDatabase();
+
+  virtual bool Open();
+
+private:
+  virtual void CreateTables();
+  virtual void CreateAnalytics();
+
+  virtual int GetSchemaVersion() const;
+  const char *GetBaseDBName() const { return "MyPrograms"; };
+};
